Set "cycle_id" in static_worker_cycle as integer

+++ grid/app/main/sfl/cycles/worker_cycle.py
@@ -19,7 +19,7 @@ class WorkerCycle(db.Model):
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     request_key = db.Column(db.String())
-    cycle_id = db.Column(db.BigInteger, db.ForeignKey("static_cycle_.id"))
+    cycle_id = db.Column(db.Integer, db.ForeignKey("static_cycle_.id"))
     worker_id = db.Column(db.String, db.ForeignKey("static_worker_.id"))
     started_at = db.Column(db.DateTime(), default=datetime.datetime.utcnow())
     is_completed = db.Column(db.Boolean(), default=False)
